Prelimiary working version
Integrated a lot of components from previous projects so that the scrapper read from CDCS then wrote the results to a psql database. Currently, there seem to be some edge cases for parsing that still fail. Also, the schema for the database is something to be desired (it is currently only one table with course information atm).